requirements.txt
google-genai>=0.1.0
litellm>=1.40.0
pydantic>=2.0.0
typer[all]>=0.9.0
rich>=13.0.0
pyyaml>=6.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
tqdm>=4.65.0
# Web UI Dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
